pyproject.toml
[project]
name = "weather-platform"
version = "0.1.0"
description = "Registry-driven ingestion + stage flatten + dbt curated transforms"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "pyyaml>=6.0.1",
  "pandas>=2.2.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "typer>=0.12.0",
  "croniter (>=6.0.0,<7.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
]

[project.scripts]
weather = "cli:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry]
name = "weather-platform"
version = "0.1.0"
description = ""
packages = [{ include = "weather_platform", from = "src" }]
[dependency-groups]
dev = [
    "dbt-postgres (>=1.10.0,<2.0.0)"
]
